+++ src/main/java/com/blueDragon/Convenience/Service/CategoryService.java
@@ -2,9 +2,13 @@ package com.blueDragon.Convenience.Service;
 
 import com.blueDragon.Convenience.Exception.CategoryInvalidValueException;
 import com.blueDragon.Convenience.Model.FoodTypeEntity;
+import com.blueDragon.Convenience.Model.Product;
 import com.blueDragon.Convenience.Repository.FoodTypeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -20,4 +24,12 @@ public class CategoryService {
             throw new CategoryInvalidValueException("Invalid category: " + category);
         }
     }
+
+    public List<String> combineFoodTypes(List<Product> products) {
+        return products.stream()
+                .flatMap(product -> product.getFoodTypes().stream()) // Assuming each Product has a List<FoodType>
+                .distinct()
+                .collect(Collectors.toList());
+    }
+
 }
