+++ src/main/java/com/blueDragon/Convenience/Service/ConvenienceService.java
@@ -1,9 +1,13 @@
 package com.blueDragon.Convenience.Service;
 
 import com.blueDragon.Convenience.Exception.CategoryInvalidValueException;
+import com.blueDragon.Convenience.Model.Product;
 import com.blueDragon.Convenience.Repository.ConvenienceEntityRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -19,4 +23,13 @@ public class ConvenienceService {
             throw new CategoryInvalidValueException("Invalid convenience: " + name);
         }
     }
+
+
+    public List<String> combineConvenienceTypes(List<Product> products) {
+        return products.stream()
+                .flatMap(product -> product.getAvailableAt().stream()) // Assuming each Product has a List<FoodType>
+                .distinct()
+                .collect(Collectors.toList());
+    }
+
 }
