+++ src/main/java/com/blueDragon/Convenience/Service/RecommendationService.java
@@ -32,6 +32,8 @@ public class RecommendationService {
     private final UserRepository userRepository;
     private final ProductLikeRepository productLikeRepository;
     private final RecommendationRepository recommendationRepository;
+    private final ConvenienceService convenienceService;
+    private final CategoryService categoryService;
     private final S3Uploader s3Uploader;
 
     @Transactional
@@ -43,8 +45,8 @@ public class RecommendationService {
 
         List<Product> productList = productService.getProductFromDto(recommendationDto.getProductList());
         Integer totalPrice = productService.sumProductPrices(productList);
-        List<String> foodTypes = productService.combineFoodTypes(productList);
-        List<String> availableAt = productService.combineConvenienceTypes(productList);
+        List<String> foodTypes = categoryService.combineFoodTypes(productList);
+        List<String> availableAt = convenienceService.combineConvenienceTypes(productList);
         List<String> urls = files.stream().map((multipartFile -> {
             try {
                 return s3Uploader.upload(multipartFile, "recommend");
